% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_utils.R
\name{compute_standardized_residuals}
\alias{compute_standardized_residuals}
\title{Compute standardized residuals for stage 1 & 2}
\usage{
compute_standardized_residuals(params, model_data, eb_estimates)
}
\arguments{
\item{params}{Vector of model parameters (beta, alpha, tau)}

\item{model_data}{List containing model matrices and data}

\item{eb_estimates}{List containing empirical Bayes estimates}
}
\value{
Vector of standardized residuals
}
\description{
Computes standardized residuals using empirical Bayes estimates.
}
