% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_params_display}
\alias{format_params_display}
\title{Format parameters for display}
\usage{
format_params_display(
  params,
  param_names,
  mean_formula,
  var_bs_formula,
  var_ws_formula,
  stage = 1,
  model_type = NULL
)
}
\arguments{
\item{params}{Vector of parameter values}

\item{param_names}{Vector of parameter names (beta_1, alpha_1, etc.)}

\item{mean_formula}{Formula for fixed effects mean model}

\item{var_bs_formula}{Formula for between-subject variance model}

\item{var_ws_formula}{Formula for within-subject variance model}

\item{stage}{stage of the model (1, 2, or 3)}

\item{model_type}{For stage 3, the type of model}
}
\description{
Format parameters for display
}
