% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gh_points}
\alias{get_gh_points}
\title{Get Gaussian-Hermite quadrature points}
\usage{
get_gh_points(n_points)
}
\arguments{
\item{n_points}{Number of quadrature points}
}
\value{
A list containing points and weights
}
\description{
Calculates Gaussian-Hermite quadrature points and weights for numerical integration
}
