% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_init.R
\name{init_1}
\alias{init_1}
\title{Initialize stage 1 parameters}
\usage{
init_1(data, fixed, bs = NULL, ws = NULL, response)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{fixed}{Formula for fixed effects mean model}

\item{bs}{Formula for between-subject variance model (or NULL)}

\item{ws}{Formula for within-subject variance model}

\item{response}{Name of response variable column}
}
\value{
List of initial parameter values and supporting information
}
\description{
Computes initial values for Stage 1 model parameters based on OLS fits
}
