% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_init.R
\name{init_2}
\alias{init_2}
\title{Initialize stage 2 parameters}
\usage{
init_2(data, response, mean_fit, tau_intercept, ws)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{response}{Name of response variable column}

\item{mean_fit}{Mean model fit from Stage 1}

\item{tau_intercept}{tau intercept parameter from Stage 1}

\item{ws}{Formula for within-subject variance model for Stage 2}
}
\value{
Vector of initial tau parameters for Stage 2
}
\description{
Computes initial values for Stage 2 model parameters based on Stage 1 results
}
