% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_utils.R
\name{prepare_model_data}
\alias{prepare_model_data}
\title{Prepare model data and matrices}
\usage{
prepare_model_data(data, fixed = NULL, bs, ws, id, response, nq)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{fixed}{Formula for fixed effects mean model}

\item{bs}{Formula for between-subject variance model}

\item{ws}{Formula for within-subject variance model}

\item{id}{Name of ID variable column}

\item{response}{Name of response variable column}

\item{nq}{Number of quadrature points}
}
\value{
A list containing all prepared data and matrices for model fitting
}
\description{
A utility function to prepare all necessary data and matrices for model fitting, avoiding repeated computations
}
