% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_method.R
\name{print.mels}
\alias{print.mels}
\title{Print method for mels objects}
\usage{
\method{print}{mels}(x, ...)
}
\arguments{
\item{x}{A mels object}

\item{...}{Additional arguments}
}
\description{
Print method for mels objects
}
