% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.mixregls}
\alias{print.mixregls}
\title{Print method for mixregls objects}
\usage{
\method{print}{mixregls}(x, ...)
}
\arguments{
\item{x}{A mixregls object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for mixregls objects
}
