% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_method.R
\name{print.summary.mels}
\alias{print.summary.mels}
\title{Print method for summary.mels objects}
\usage{
\method{print}{summary.mels}(x, ...)
}
\arguments{
\item{x}{A summary.mels object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for summary.mels objects
}
