% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{print.summary.mixregls}
\alias{print.summary.mixregls}
\title{Print method for summary.mixregls objects}
\usage{
\method{print}{summary.mixregls}(x, ...)
}
\arguments{
\item{x}{A summary.mixregls object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for summary.mixregls objects
}
