% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{run_em}
\alias{run_em}
\title{Run EM algorithm for initial parameter estimation}
\usage{
run_em(
  data,
  matrices,
  mean_formula,
  id_var,
  response_var,
  init_values,
  max_iter = 20
)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{matrices}{Design matrices from setup_matrices}

\item{mean_formula}{Formula for fixed effects mean model}

\item{id_var}{Name of ID variable column}

\item{response_var}{Name of response variable column}

\item{init_values}{List of initial values from init_1}

\item{max_iter}{Maximum number of EM iterations}
}
\value{
Updated init_values list with refined parameter estimates
}
\description{
Refines initial parameter estimates using EM algorithm
NOTE: This function assumes a random intercept model
}
