% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_utils.R
\name{setup_matrices}
\alias{setup_matrices}
\title{Setup design matrices for model fitting}
\usage{
setup_matrices(data, fixed, bs, ws)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{fixed}{Formula for fixed effects mean model (or NULL for intercept-only)}

\item{bs}{Formula for between-subject variance model}

\item{ws}{Formula for within-subject variance model}
}
\value{
A list containing design matrices X, U, and Z
}
\description{
Creates design matrices for fixed effects, between-subject variance, and within-subject variance
}
