% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stage12_gradient_rcpp}
\alias{stage12_gradient_rcpp}
\title{Calculate gradient for Stage 1 & 2 log-likelihood (Fixed Quadrature)}
\usage{
stage12_gradient_rcpp(params, X, U, Z, y, id, points, weights)
}
\arguments{
\item{params}{A vector of model parameters (beta, alpha, tau).}

\item{X}{Design matrix for fixed effects.}

\item{U}{Design matrix for between-subject variance.}

\item{Z}{Design matrix for within-subject variance.}

\item{y}{The response vector.}

\item{id}{A vector of subject IDs.}

\item{points}{Vector of standard Quadrature points.}

\item{weights}{Vector of standard Quadrature weights.}
}
\value{
A vector containing the gradient of the log-likelihood.
}
\description{
Computes the gradient of the log-likelihood for the mixed effects location-scale model
(Stage 1 & 2) using standard Gaussian quadrature.
}
\keyword{internal}
