% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_loglike.R
\name{stage12_hessian_wrapper}
\alias{stage12_hessian_wrapper}
\title{Stage 1 & 2 hessian function}
\usage{
stage12_hessian_wrapper(
  params,
  model_data,
  adaptive = 1,
  eb_theta = NULL,
  eb_sd = NULL
)
}
\arguments{
\item{params}{Vector of parameters (beta, alpha, tau)}

\item{model_data}{Prepared model data from prepare_model_data function}

\item{adaptive}{Use adaptive quadrature (0 = fixed, 1 = adaptive)}

\item{eb_theta}{Empirical Bayes estimates for each subject (only used if adaptive = 1)}

\item{eb_sd}{Posterior standard deviations for each subject (only used if adaptive = 1)}
}
\value{
Hessian matrix
}
\description{
Wrapper for C++ Hessian function, supporting adaptive quadrature
}
