% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_method.R
\name{summary.mels}
\alias{summary.mels}
\title{Summary method for mels objects}
\usage{
\method{summary}{mels}(object, ...)
}
\arguments{
\item{object}{A mels object}

\item{...}{Additional arguments}
}
\description{
Summary method for mels objects
}
