% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{summary.mixregls}
\alias{summary.mixregls}
\title{Summary method for mixregls objects}
\usage{
\method{summary}{mixregls}(object, ...)
}
\arguments{
\item{object}{A mixregls object}

\item{...}{Additional arguments (not used)}
}
\description{
Summary method for mixregls objects
}
