% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_utils.R
\name{update_eb_estimates}
\alias{update_eb_estimates}
\title{Update Empirical Bayes estimates for stage 1 & 2}
\usage{
update_eb_estimates(params, model_data, eb_prev_theta, eb_prev_psd)
}
\arguments{
\item{params}{Vector of current model parameters (beta, alpha, tau).}

\item{model_data}{List containing model matrices and data.}

\item{eb_prev_theta}{EB mean estimates from the previous iteration.}

\item{eb_prev_psd}{EB standard deviation estimates from the previous iteration.}
}
\value{
A list containing the newly calculated \code{eb_theta} and \code{eb_sd}.
}
\description{
This function calls C++ function to calculate the EB estimates.
}
