% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mels_utils.R
\name{update_eb_estimates_stage3}
\alias{update_eb_estimates_stage3}
\title{Update Bivariate Empirical Bayes estimates for stage 3}
\usage{
update_eb_estimates_stage3(
  params,
  model_data,
  model_type,
  eb_prev_theta2,
  eb_prev_thetav,
  adaptive
)
}
\arguments{
\item{params}{Vector of current model parameters.}

\item{model_data}{List containing model matrices and data.}

\item{model_type}{The specific stage 3 model type.}

\item{eb_prev_theta2}{A matrix of the posterior means from the previous iteration.}

\item{eb_prev_thetav}{A matrix of the posterior VCV elements from the previous iteration.}

\item{adaptive}{An integer (0 or 1) indicating whether to use adaptive quadrature.}
}
\value{
A list containing the newly calculated \code{theta2_eb} and \code{thetav_eb}.
}
\description{
Update Bivariate Empirical Bayes estimates for stage 3
}
