% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{init_1}
\alias{init_1}
\title{Initialize stage 1 parameters}
\usage{
init_1(
  data,
  mean_formula,
  var_bs_formula = NULL,
  var_ws_formula = NULL,
  response_var
)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{mean_formula}{Formula for fixed effects mean model}

\item{var_bs_formula}{Formula for between-subject variance model (or NULL)}

\item{var_ws_formula}{Formula for within-subject variance model}

\item{response_var}{Name of response variable column}
}
\value{
List of initial parameter values and supporting information
}
\description{
Computes initial values for Stage 1 model parameters based on OLS fits
}
