% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_model_data}
\alias{prepare_model_data}
\title{Prepare model data and matrices}
\usage{
prepare_model_data(
  data,
  mean_formula = NULL,
  var_bs_formula,
  var_ws_formula,
  id_var,
  response_var,
  nq
)
}
\arguments{
\item{data}{Data frame containing model variables}

\item{mean_formula}{Formula for fixed effects mean model}

\item{var_bs_formula}{Formula for between-subject variance model}

\item{var_ws_formula}{Formula for within-subject variance model}

\item{id_var}{Name of ID variable column}

\item{response_var}{Name of response variable column}

\item{nq}{Number of quadrature points}
}
\value{
A list containing all prepared data and matrices for model fitting
}
\description{
A utility function to prepare all necessary data and matrices for model fitting, avoiding repeated computations
}
