% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stage12_gradient_adaptive}
\alias{stage12_gradient_adaptive}
\title{Calculate gradient for Stage 1 & 2 log-likelihood (Adaptive Quadrature)}
\usage{
stage12_gradient_adaptive(
  params,
  X,
  U,
  Z,
  y,
  id,
  points,
  weights,
  adaptive,
  eb_theta,
  eb_sd
)
}
\arguments{
\item{params}{A vector of model parameters (beta, alpha, tau).}

\item{X}{Design matrix for fixed effects.}

\item{U}{Design matrix for between-subject variance.}

\item{Z}{Design matrix for within-subject variance.}

\item{y}{The response vector.}

\item{id}{A vector of subject IDs.}

\item{points}{Vector of standard Quadrature points.}

\item{weights}{Vector of standard Quadrature weights.}

\item{adaptive}{Integer flag (0 or 1) indicating if adaptive quadrature is used.}

\item{eb_theta}{Vector of Empirical Bayes estimates for random effect means.}

\item{eb_sd}{Vector of Empirical Bayes estimates for random effect standard deviations.}
}
\value{
A vector containing the gradient of the log-likelihood.
}
\description{
Computes the gradient of the log-likelihood for the mixed effects location-scale model
(Stage 1 & 2) using adaptive Gaussian quadrature.
}
\keyword{internal}
