% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stage12_hessian_rcpp}
\alias{stage12_hessian_rcpp}
\title{Calculate Hessian matrix for Stage 1 & 2 log-likelihood (C++)}
\usage{
stage12_hessian_rcpp(params, X, U, Z, y, id, points, weights)
}
\arguments{
\item{params}{Vector of parameters (beta, alpha, tau).}

\item{X}{Design matrix for fixed effects.}

\item{U}{Design matrix for between-subject variance.}

\item{Z}{Design matrix for within-subject variance.}

\item{y}{Vector of response values.}

\item{id}{Vector of subject identifiers (0-based).}

\item{points}{Standard Gaussian-Hermite quadrature points.}

\item{weights}{Standard Gaussian-Hermite quadrature weights.}
}
\value{
The Hessian matrix of the log-likelihood.
}
\description{
This function computes the Hessian matrix (second derivatives) of the
log-likelihood for the Stage 1 & 2 models using fixed Gaussian-Hermite quadrature.
}
