% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_eb_estimates_rcpp}
\alias{update_eb_estimates_rcpp}
\title{Update Empirical Bayes Estimates for Stage 1 & 2}
\usage{
update_eb_estimates_rcpp(
  params,
  X,
  U,
  Z,
  y,
  id,
  points,
  weights,
  eb_prev_theta,
  eb_prev_psd
)
}
\arguments{
\item{params}{A vector of model parameters (beta, alpha, tau).}

\item{X}{Design matrix for fixed effects.}

\item{U}{Design matrix for between-subject variance.}

\item{Z}{Design matrix for within-subject variance.}

\item{y}{The response vector.}

\item{id}{A vector of subject IDs.}

\item{points}{Vector of standard Quadrature points.}

\item{weights}{Vector of standard Quadrature weights.}

\item{eb_prev_theta}{Vector of previous iteration's EB mean estimates.}

\item{eb_prev_psd}{Vector of previous iteration's EB standard deviation estimates.}
}
\value{
A list containing:
\item{eb_theta}{Updated vector of random effect means.}
\item{eb_psd}{Updated vector of random effect standard deviations.}
}
\description{
Calculates the updated Empirical Bayes (EB) estimates for the random effects
means (theta) and standard deviations (psd) using adaptive Gaussian quadrature.
}
\keyword{internal}
